move argparse and logging setup to get_args module
